migrate from old to new endpoints
